check_f2fs: fix build error in abfarm474
This patch fixes this build error:

error: comparison of integers of different signs: 'long long' and 'unsigned long long' [-Werror,-Wsign-compare]

if (sb.st_size / 512 != sb.st_blocks) {
~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~
1 error generated.

Bug: 77146255
Bug: 77344057
Fixes: 160ec1fef58d ("check_f2fs: test wrong write(2) call")

+++ tools/check_f2fs.c
@@ -114,7 +114,7 @@ static int test_bad_write_call(char *path)
 		return -1;
 	}
 
-	if (sb.st_size / 512 != sb.st_blocks) {
+	if ((long long)sb.st_size / 512 != (long long)sb.st_blocks) {
 		printf("FAIL: Mismatch i_size and i_blocks: %lld %lld\n",
 			(long long)sb.st_size, (long long)sb.st_blocks);
 		printf("FAIL: missing patch "
